Return 401 instead of 403 when the tokens are bad.
In `smart/middlewares/authentication` we check for the presence of an OAuth header, and if it's there but no principal comes out of it (i.e. the header is invalid), a 401 is returned. There was no check in place there before, so only when `smart/middlewares/authorization` was hit it would return a 403. That's because the permission to request a token was not present if there was no (valid) app principal in the request.